refactor: centralize plan types
