+++ src/types/index.ts
@@ -39,87 +39,6 @@ export interface UserContext {
     teams: TeamMembershipSummary[];
 }
 
-// Base Plan type (list view)
-export interface Plan {
-    id: string;
-    title: string;
-    description?: string | null;
-    difficulty_level?: number | null;
-    duration_weeks?: number | null;
-    sport?: string | null;
-    created_by?: string | null;
-    team_id?: string | null;
-    visibility: 'public' | 'team' | 'private';
-    allow_public_forking: boolean;
-    origin_type: 'user' | 'admin' | 'ai' | 'coach';
-    is_featured: boolean;
-    view_count: number;
-    fork_count: number;
-    like_count: number;
-    forked_from?: string | null;
-    created_at: string; // ISO Date String
-    updated_at?: string | null;
-}
-
-// Type for discoverable plans RPC (includes creator/team names)
-export interface DiscoverablePlan extends Plan {
-     team_name?: string | null;
-     created_by_username?: string | null;
-     created_by_profile_image_url?: string | null;
-}
-
-// Type for Exercise (basic info)
-export interface ExerciseSummary {
-    id: string;
-    name: string;
-    image_url?: string | null;
-    video_url?: string | null;
-}
-
-// Type for Plan Exercise Detail
-export interface PlanExerciseDetail {
-    id: string;
-    plan_day_id: string;
-    exercise_id: string;
-    order_index: number;
-    sets?: string | null;
-    reps?: string | null;
-    duration?: string | null;
-    distance?: string | null;
-    rest_time?: string | null;
-    target_sets_min?: number | null;
-    target_sets_max?: number | null;
-    // ... other target fields ...
-    notes?: string | null;
-    exercises: ExerciseSummary; // Nested Exercise info
-}
-
-// Type for Plan Day Detail
-export interface PlanDayDetail {
-    id: string;
-    plan_week_id: string;
-    day_number: number;
-    title?: string | null;
-    description?: string | null;
-    is_rest_day: boolean;
-    plan_exercises: PlanExerciseDetail[]; // Nested exercises
-}
-
-// Type for Plan Week Detail
-export interface PlanWeekDetail {
-    id: string;
-    plan_id: string;
-    week_number: number;
-    description?: string | null;
-    plan_days: PlanDayDetail[]; // Nested days
-}
-
-// Type for Full Plan Detail (used in fetchPlanDetails)
-export interface PlanDetail extends Plan {
-    profiles?: Pick<UserProfile, 'username' | 'profile_image_url'> | null; // Creator profile subset
-    teams?: Pick<Team, 'name' | 'logo_url'> | null; // Team subset
-    plan_weeks: PlanWeekDetail[]; // Fully nested structure
-}
 
 // Type for Set Log
 export interface SetLog {
